// src/app/enums/enum-race.ts
export enum RaceName {
  // DEFAULT
  NULL = '',
  // Standard Races PHB
  HUMAN = 'Human',
  ELF = 'Elf',
  HALF_ELF = 'Half Elf',
  DWARF = 'Dwarf',
  HALFLING = 'Halfling',
  GNOME = 'Gnome',
  HALF_ORC = 'Half Orc',
  // EXTRA
  AASIMAR = 'Aasimar',
  WARFORGED = 'Warforged',
  GOBLIN = 'Goblin',
  KOBOLD = 'Kobold',
  ORC = 'Orc',
  TIEFLING = 'Tiefling',
  TIBBIT = 'Tibbit'
}